+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Board Support Library Changelog
   * ADDED: Callable from C
   * ADDED: I2C master exit API for XK-AUDIO-316-MC
   * FIXED: Uninitialised global interface for XK-AUDIO-216-MC setup
+  * ADDED: Explicit xk_evk_xu316_AudioHwChanInit() API
 
 0.1.1
 -----