+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-Board Support Library Changelog
-===============================
+lib_board_support change log
+============================
 
 1.0.0
 -----
@@ -12,6 +12,14 @@ Board Support Library Changelog
   * FIXED: Uninitialised global interface for XK-AUDIO-216-MC setup
   * ADDED: Explicit xk_evk_xu316_AudioHwChanInit() API
 
+  * Changes to dependencies:
+
+    - lib_i2c: Added dependency 6.2.0
+
+    - lib_sw_pll: Added dependency 2.2.0
+
+    - lib_xassert: Added dependency 4.2.0
+
 0.1.1
 -----
 
@@ -20,4 +28,5 @@ Board Support Library Changelog
 0.1.0
 -----
 
-  * ADDED: Ported setup code from sw_usb_audio v8.1.0 
+  * ADDED: Ported setup code from sw_usb_audio v8.1.0
+